Add files via upload
This project is a web-based application built using Flask that allows users to upload brain MRI images and get predictions on whether the MRI indicates the presence of a brain tumor or not. The backend uses a VGG16 deep learning model, trained on a dataset of MRI images, to classify the images into two categories: Tumor and No Tumor.